Handle nullability and nesting (#20)
* Handle nullability and nesting

this PR was supposed to handle just the Nullability part, but turned out
to be much bigger since once I started to implement that, I started
bumping into roadblocks with the current implementation that wouldn't
let me properly propragate the nullability information without
refactoring a lot of code.
It ended up being much easier to also make the change to generated
nested types with the correct names based on the responseName field.
The code turned out to be **at lot** simpler and much more similar to
the Swift counterparts. This still doesnt handle deduping of duplicate
types but that shouldnt be too hard to implement now. We just need to
decide on how to actualy do it (maybe add a suffix to the interface
name? Maybe a number, or a letter, or a symbol?) I'm open to ideas here.

* Code review nits