add apollo-codgen gradle tasks (#52)
* add apollo-codgen gradle tasks

This adds an npm install task for apollo-codgen and refactors some of the gradle
tasks.
This also adds an IR gen task which runs apollo-codgen on all .graphql files under a
variant's sourceSet, generating a single variantNameAPI.json IR file

* add class generator task

The task calls GraphQLCompiler and depends on the IR task. Inputs should have
the format "**API.json" and should exist in the same output dir as the IR gen
task. For instance, 'src/main/graphql/generatedIR/MainAPI.json'

* add schemaPath to extensions, refactor the schema and query file lookup

Also, addresses the review comments

* add basic project setup tests

The tests ensure that the correct setup for the tasks happen, outputs
directories created and task dependencies are as expected.
Those don't test the execution behaviour as this would require separate
integration tests