Bump Apollo Codgen to version 0.10.8. Simplify compiler test fixture … (#231)
* Bump Apollo Codgen to version 0.10.8. Simplify compiler test fixture generation.

Apollo-codegen 0.10.8 supports field arguments, which are needed to generate
cache keys.

The process of updating the code-gen version was very manual as the
apollo code-gen tests run from pure IR (TestQuery.json) files.

When we bump the codegen, we will want to regenerate the TestQuery.json
files. This diff extracts the TestQuery.graphql for each of the compiler
tests and creates a simple gradle plugin to execute apollo-codegen
with the *Local* version of apollo-codegen.

While it would be nice to have this happen automatically, doing so
would require following a appraoch similiar to the apollo-gradle-plugin
which installs apollo-codegen to a local repository. Sharing this code
would require a more significant refactor to avoid logic duplication.

Instead, whenever we update the codegen plugin, we can run
`./gradlew :apollo-compiler:updateTestIRFiles`.

This diff also tried to unify the compiler tests on one
schema. Custom logic needed for certain tests was injected
into the schema. In cases where the test needs something completely
custom, you can place a schema.json file in the test folder.

* make sure operation name matches expected file name

* remove unnecessary import

* move apollocodegen version to compiler companion object